refactor: move complexity constants to new file
